Accessibility Improvements
Improvements to accessibility so that screen readers can interpret component contents, added alt text to image that uses Value property and component SecondaryText Input property. Gave component image a TabIndex of zero so that it can be accessed by screen readers. Gave component image an OnSelect that creates a Notify of the alt text (tooltip). Renamed controls and screens in example App to a consistent standard.